requirements.txt
requests~=2.20.0
websockets~=8.1
pyserial~=3.4
